bot/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
lxml==6.0.2
multidict==6.7.0
osrparse==7.0.1
pillow==12.0.0
propcache==0.4.1
pydub==0.25.1
python-telegram-bot==22.5
requests==2.32.5
rosu-pp-py==3.1.0
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0
